\name{MBESS}
\alias{MBESS}
\alias{mbess}
\alias{MBES}
\alias{mbes}
\docType{package}
\title{MBESS}
\description{
\code{MBESS} implements methods that are especially applicable to 
researchers working within the behavioral, educational, and social
sciences. Many of the functions are also applicable to disciplines related to the behavioral,
educational, and social sciences.
}
\details{
\tabular{ll}{
Package: \tab MBESS\cr
Type: \tab Package\cr
Version: \tab 3.2.0 \cr
Date: \tab 2010-11-21\cr
License: \tab GPL(>=2)\cr
}
Please read the manual and visit the corresponding web site 
\url{http://nd.edu/~kkelley/site/MBESS.html} 
for information on the capabilities of the \code{MBESS} package. Feel free 
to contact me if there is a feature you would like to see added if it would 
complement the goals of the MBESS package.
}
\author{
Ken Kelley <\email{KKelley@ND.Edu}; http://www.nd.edu/~kkelley> and Keke Lai <Lai.15@ND.Edu>

Maintainer: Keke Lai <Lai.15@ND.Edu>; 
Ken Kelley <\email{KKelley@ND.Edu}; http://www.nd.edu/~kkelley>
}

\keyword{package}
