\name{MBESS}
\alias{MBESS}
\alias{mbess}
\alias{MBES}
\alias{mbes}
\docType{package}
\title{Methods for the Behavioral, Educational, and Social Sciences}
\description{
\code{MBESS} implements methods that are especially applicable to 
researchers working within the behavioral, educational, and social
sciences. Many of the functions are also applicable to disciplines related to the behavioral,
educational, and social sciences.
}
\details{
\tabular{ll}{
Package: \tab MBESS\cr
Type: \tab Package\cr
Version: \tab 0.0.2 \cr
Date: \tab 2006-05-9\cr
License: \tab General Public License 2 or more recent.\cr
}
Please read the manual and visit the corresponding web site 
\url{http://www.indiana.edu/~kenkel/mbess} for information on the capabilities
of the \code{MBESS} package. Feel free to contact me if there is a feature you
would like to see added if it would complement the goals of the package.
}
\author{Ken Kelley <\email{KKIII@Indiana.Edu}; http://www.indiana.edu/~kenkel>

Maintainer: Ken Kelley <\email{KKIII@Indiana.Edu}; http://www.indiana.edu/~kenkel/mbess>
}

\keyword{package}
